Use trusty backports; not Debian sid
Travis's stuff is... odd. They white list a bajillion external repos, but not
Canonical's official backports for Trusty. One of many unfortunate oddities on
the service.

Because it's a new repo, I have to use "require: sudo", which gets us on the VM
infra rather than the container infra. It's worth it though, to actually get
this working.